move headers, cookies and connection point to JettyApplicationRequest
