+++ ktor-server/ktor-server-jetty-jakarta/jvm/src/io/ktor/server/jetty/jakarta/JettyApplicationRequest.kt
@@ -13,6 +13,7 @@ import io.ktor.utils.io.jvm.javaio.*
 import org.eclipse.jetty.io.*
 import org.eclipse.jetty.server.*
 
+@InternalAPI
 public class JettyApplicationRequest(
     call: PipelineCall,
     request: Request
@@ -32,4 +33,73 @@ public class JettyApplicationRequest(
         val uri = request.httpURI.query ?: return@lazy Parameters.Empty
         parseQueryString(uri, decode = false)
     }
+}
+
+@InternalAPI
+public class JettyRequestCookies(
+    request: JettyApplicationRequest,
+    private val jettyRequest: Request
+) : RequestCookies(request) {
+    override fun fetchCookies(): Map<String, String> {
+        return Request.getCookies(jettyRequest).associate { it.name to it.value }
+    }
+}
+
+@InternalAPI
+public class JettyHeaders(
+    private val jettyRequest: Request
+) : Headers {
+    override val caseInsensitiveName: Boolean = true
+
+    override fun entries(): Set<Map.Entry<String, List<String>>> {
+        return jettyRequest.headers.fieldNamesCollection.map {
+            object : Map.Entry<String, List<String>> {
+                override val key: String = it
+                override val value: List<String> = jettyRequest.headers.getValuesList(it)
+            }
+        }.toSet()
+    }
+
+    override fun getAll(name: String): List<String>? = jettyRequest.headers.getValuesList(name)
+
+    override fun get(name: String): String? = jettyRequest.headers.get(name)
+
+    override fun isEmpty(): Boolean = jettyRequest.headers.size() == 0
+
+    override fun names(): Set<String> = jettyRequest.headers.fieldNamesCollection
+}
+
+@InternalAPI
+public class JettyConnectionPoint(
+    request: Request
+) : RequestConnectionPoint {
+    @Deprecated("Use localHost or serverHost instead")
+    override val host: String = request.httpURI.host
+
+    override val localAddress: String = Request.getLocalAddr(request)
+
+    override val localHost: String = Request.getServerName(request)
+
+    override val localPort: Int = Request.getLocalPort(request)
+
+    override val method: HttpMethod = HttpMethod.parse(request.method)
+
+    @Deprecated("Use localPort or serverPort instead")
+    override val port: Int = request.httpURI.port
+
+    override val remoteAddress: String = Request.getRemoteAddr(request)
+
+    override val remoteHost: String = Request.getServerName(request)
+
+    override val remotePort: Int = Request.getRemotePort(request)
+
+    override val scheme: String = request.httpURI.scheme
+
+    override val serverHost: String = Request.getServerName(request)
+
+    override val serverPort: Int = Request.getServerPort(request)
+
+    override val uri: String = request.httpURI.pathQuery
+
+    override val version: String = request.connectionMetaData.httpVersion.asString()
 }
